Add external config file for config changes after compilation; try to stabilize working cycle execution time (still far from perfect); use context-relevant explanations; only increase concept priority when processing events; only construct temporal implications for precondition and operation; use working cycles as interval; bug fixes;

+++ NARSDataStructures/Bag.py
@@ -165,7 +165,7 @@ class Bag(ItemContainer):
         # pop random item from currently selected bucket
         rnd = random.random()  # another randomly generated number in [0.0, 1.0)
         maxidx = len(self.buckets[self.current_bucket_number]) - 1
-        randidx = int(round(rnd * maxidx))
+        randidx = round(rnd * maxidx)
         item = self.buckets[self.current_bucket_number][randidx]
         return item, randidx
